IC2 energynet compat (#21)
* Begin adding support for IC2 energy

This is still extremely broken and contains debug code!

* Serverside integration mostly finished.

This means all that's left is clientside rendering. Fun times.

* Finished clientside rendering.

A limitation of the current implementation is that up to 80 ticks must elapse, along with a block update occuring.

While none of these are absolutely breaking, I may try and fix them in the future. Until then, this should be regarded as functional.